feat: query vo moved to query directory
