+++ backend/app/Recommendation/Domain/Model/ValueObjects/Query/Title.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Recommendation\Domain\Model\ValueObjects;
+namespace App\Recommendation\Domain\Model\ValueObjects\Query;
 
 use Exception;
